Update README.md
Hey! Welcome to my project, NBA Player Analysis. The intent of this exercise is to explore, clean, and analyze data on NBA Players to uncover insights from the following perspectives:

As an NBA Team Owner, I would like to know which players have the highest PER so that I know who to target in free agency to provide the most excitement to my fans.

As an NBA Head Coach, I would like to know players' win/loss percentages so that I can target players with a winning pedigree in free agency.

As an NBA Player, I would like to know coaches with high PPG numbers so that I can consider who can help me unlock my offense.